+++ Runtime/ISaveService.cs
@@ -6,7 +6,7 @@ namespace DTech.DataPersistence
     {
         bool HasSave(string key);
         Task SaveAsync<T>(string key, T value, bool isCrypted = true);
-        Task<T> LoadAsync<T>(string key, T defaultValue);
+        Task<T> LoadAsync<T>(string key, T defaultValue, bool isCrypted = true);
         void Remove(string key);
     }
 }