+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.1] - 2025-10-20
+
+### Added
+- Internal ``SaveEntity``
+
+### Fixed
+- Fixed issue with always save crypted value
+
 ## [0.1.0] - 2025-10-13
 
 ### Added